solve errors regrding paths

+++ routes/review.js
@@ -1,15 +1,13 @@
 const express = require("express");
-const router = express.Router({mergeParams: true});
+const router = express.Router({ mergeParams: true });
 const wrapAsync = require("../utils/wrapAsync.js");
 const Review = require("../models/review.js");
 const Listing = require("../models/listing.js");
 const { validateReview } = require("../middlewares/validateReview.js");
 const { isLoggedIn, isReviewAuthor } = require("../middlewares/authenticate.js");
 
-
-
-//Reviews Post route
-router.post("/",isLoggedIn,validateReview,wrapAsync(async(req,res)=> {
+// Post route for creating a review
+router.post("/", isLoggedIn, validateReview, wrapAsync(async (req, res) => {
   let listing = await Listing.findById(req.params.id);
   let newReview = new Review(req.body.review);
   newReview.author = req.user._id;
@@ -17,19 +15,23 @@ router.post("/",isLoggedIn,validateReview,wrapAsync(async(req,res)=> {
   await newReview.save();
   await listing.save();
 
-  // console.log("new review saved");
-  // res.send("new review saved");
-
   req.flash("success", "New review created!");
   res.redirect(`/listings/${listing._id}`);
 }));
 
-//Delete Route of reviews
-router.delete("/:reviewId",isLoggedIn,isReviewAuthor,wrapAsync(async(req,res)=> {
+// Delete route for deleting a review
+router.delete("/:reviewId", isLoggedIn, isReviewAuthor, wrapAsync(async (req, res) => {
   const { id, reviewId } = req.params;
 
-  await Listing.findByIdAndUpdate(id,{$pull: {reviews: reviewId}});
+  const listing = await Listing.findById(id);
+  if (!listing) {
+    req.flash("error", "Listing not found!");
+    return res.redirect("/listings");
+  }
+
   await Review.findByIdAndDelete(reviewId);
+  await Listing.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
+
   req.flash("success", "Review deleted!");
   res.redirect(`/listings/${id}`);
 }));
